refactor: cleanup for review
