+++ packages/slice-machine/src/features/customTypes/customTypesBuilder/ImportSlicesFromLibraryModal/LibrarySlicesTab.tsx
@@ -38,7 +38,7 @@ export function LibrarySlicesTab() {
               onValueChange={setGithubUrl}
             />
             <Button
-              onClick={() => handleImportFromGithub(githubUrl)}
+              onClick={() => void handleImportFromGithub(githubUrl)}
               disabled={!githubUrl.trim() || isLoadingSlices}
               loading={isLoadingSlices}
               color="purple"
@@ -79,4 +79,3 @@ export function LibrarySlicesTab() {
     </ScrollArea>
   );
 }
-
