--- packages/slice-machine/src/features/customTypes/customTypesBuilder/ReuseExistingSlicesDialog/index.tsx
@@ -1,1 +1,0 @@
-export { ReuseExistingSlicesDialog } from "./ReuseExistingSlicesDialog";
