--- packages/slice-machine/src/features/customTypes/customTypesBuilder/ReuseExistingSlicesDialog/utils/sliceWithoutConflicts.ts
@@ -1,51 +1,0 @@
-import { SharedSlice } from "@prismicio/types-internal/lib/customtypes";
-
-import { NewSlice } from "../types";
-
-/**
- * If needed, assigns new ids and names to avoid conflicts with existing slices.
- * Names are compared case-insensitively to avoid conflicts
- * between folder names with different casing.
- */
-export function sliceWithoutConflicts({
-  existingSlices,
-  newSlices,
-  slice,
-}: {
-  existingSlices: SharedSlice[];
-  newSlices: NewSlice[];
-  slice: SharedSlice;
-}): SharedSlice {
-  const existingIds = new Set<string>();
-  const existingNames = new Set<string>();
-
-  for (const { id, name } of existingSlices) {
-    existingIds.add(id);
-    existingNames.add(name.toLowerCase());
-  }
-
-  for (const s of newSlices) {
-    existingIds.add(s.model.id);
-    existingNames.add(s.model.name.toLowerCase());
-  }
-
-  let id = slice.id;
-  let counter = 2;
-  while (existingIds.has(id)) {
-    id = `${slice.id}_${counter}`;
-    counter++;
-  }
-
-  let name = slice.name;
-  counter = 2;
-  while (existingNames.has(name.toLowerCase())) {
-    name = `${slice.name}${counter}`;
-    counter++;
-  }
-
-  return {
-    ...slice,
-    id,
-    name,
-  };
-}
